requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
httpx>=0.27
openai>=1.30
anthropic>=0.25
google-generativeai>=0.5
beautifulsoup4>=4.12
rich>=13.7
pyyaml>=6.0
tiktoken>=0.7
fastapi>=0.110
uvicorn>=0.29
python-multipart>=0.0.9
google-auth>=2.29
google-auth-oauthlib>=1.2
google-api-python-client>=2.130
ocrmypdf>=16.0
Pillow>=10.0
pdf2image>=1.17
spacy>=3.7
sentence-transformers>=3.0
textstat>=0.7
pymupdf>=1.24
python-docx>=1.1
numpy>=1.26
cryptography>=42.0
